Rename ByteVectorScorer* to Int8VectorScorer* and FloatVectorScorer* to Float32VectorScorer*
