+++ libs/simdvec/src/test/java/org/elasticsearch/simdvec/Int8VectorScorerFactoryTests.java
@@ -38,7 +38,7 @@ import static org.elasticsearch.simdvec.VectorSimilarityType.MAXIMUM_INNER_PRODU
 import static org.elasticsearch.simdvec.internal.vectorization.JdkFeatures.SUPPORTS_HEAP_SEGMENTS;
 import static org.hamcrest.Matchers.closeTo;
 
-public class ByteVectorScorerFactoryTests extends AbstractVectorTestCase {
+public class Int8VectorScorerFactoryTests extends AbstractVectorTestCase {
 
     private static final double DELTA = 1e-6;
 
@@ -106,7 +106,7 @@ public class ByteVectorScorerFactoryTests extends AbstractVectorTestCase {
                     var values = vectorValues(dims, size, in, sim.function());
                     float expected = luceneScore(sim, vectors[idx0], vectors[idx1]);
 
-                    var supplier = factory.getByteVectorScorerSupplier(sim, in, values).get();
+                    var supplier = factory.getInt8VectorScorerSupplier(sim, in, values).get();
                     var scorer = supplier.scorer();
                     scorer.setScoringOrdinal(idx0);
 
@@ -144,7 +144,7 @@ public class ByteVectorScorerFactoryTests extends AbstractVectorTestCase {
                         var values = vectorValues(dims, size, in, sim.function());
                         float expected = luceneScore(sim, vector(idx0, dims), vector(idx1, dims));
 
-                        var supplier = factory.getByteVectorScorerSupplier(sim, in, values).get();
+                        var supplier = factory.getInt8VectorScorerSupplier(sim, in, values).get();
                         var scorer = supplier.scorer();
                         scorer.setScoringOrdinal(idx0);
 
@@ -200,7 +200,7 @@ public class ByteVectorScorerFactoryTests extends AbstractVectorTestCase {
                     for (int i = 0; i < size; i++) {
                         expected[i] = luceneScore(sim, vectors[idx0], vectors[nodes[i]]);
                     }
-                    var supplier = factory.getByteVectorScorerSupplier(sim, in, values).get();
+                    var supplier = factory.getInt8VectorScorerSupplier(sim, in, values).get();
                     var scorer = supplier.scorer();
                     scorer.setScoringOrdinal(idx0);
                     scorer.bulkScore(nodes, scores, nodes.length);
@@ -215,7 +215,7 @@ public class ByteVectorScorerFactoryTests extends AbstractVectorTestCase {
         }
     }
 
-    // -- Query-side scorer tests (ByteVectorScorer via getByteVectorScorer, JDK 22+) --
+    // -- Query-side scorer tests (Int8VectorScorer via getInt8VectorScorer, JDK 22+) --
     // These test the query scorer which accepts both MMap and DirectAccessInput (SNAP).
 
     public void testScorerWithMMap() throws IOException {
@@ -256,7 +256,7 @@ public class ByteVectorScorerFactoryTests extends AbstractVectorTestCase {
                 for (var sim : List.of(DOT_PRODUCT, EUCLIDEAN, COSINE, MAXIMUM_INNER_PRODUCT)) {
                     var values = vectorValues(dims, size, in, sim.function());
                     float expected = luceneScore(sim, queryVector, vectors[idx]);
-                    var scorer = factory.getByteVectorScorer(sim.function(), values, queryVector).get();
+                    var scorer = factory.getInt8VectorScorer(sim.function(), values, queryVector).get();
                     double expectedDelta = Math.max(Math.abs(expected) * DELTA, DELTA);
                     assertThat(sim.toString(), (double) scorer.score(idx), closeTo(expected, expectedDelta));
                 }
@@ -317,7 +317,7 @@ public class ByteVectorScorerFactoryTests extends AbstractVectorTestCase {
                     for (int i = 0; i < size; i++) {
                         expected[i] = luceneScore(sim, queryVector, vectors[nodes[i]]);
                     }
-                    var scorer = factory.getByteVectorScorer(sim.function(), values, queryVector).get();
+                    var scorer = factory.getInt8VectorScorer(sim.function(), values, queryVector).get();
                     scorer.bulkScore(nodes, scores, nodes.length);
                     for (int i = 0; i < size; i++) {
                         double expectedDelta = Math.max(Math.abs(expected[i]) * DELTA, DELTA);
@@ -352,7 +352,7 @@ public class ByteVectorScorerFactoryTests extends AbstractVectorTestCase {
             try (IndexInput in = dir.openInput(fileName, IOContext.DEFAULT)) {
                 for (var sim : List.of(DOT_PRODUCT, EUCLIDEAN, COSINE, MAXIMUM_INNER_PRODUCT)) {
                     var values = vectorValues(dims, size, in, sim.function());
-                    var scorer = factory.getByteVectorScorer(sim.function(), values, queryVector).get();
+                    var scorer = factory.getInt8VectorScorer(sim.function(), values, queryVector).get();
                     float result = scorer.bulkScore(new int[0], new float[0], 0);
                     assertEquals(Float.NEGATIVE_INFINITY, result, 0f);
                 }
